Labs/Lab_2/Airflow_labs/requirements.txt
apache-airflow>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
